+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 ## 1.0.4
 - Add getAccounts command
+- Add addService command
+
 ## 1.0.3
 - Add macOS configuration in README.md